perf: helpline code
