+++ docSite/content/zh-cn/docs/development/upgrading/4810.md
@@ -20,4 +20,6 @@ weight: 816
 ## V4.8.10 更新说明
 
 1. 新增 - 模板市场
-2. 
+2. 新增 - 工作流节点拖动自动对齐吸附
+3. 商业版新增 - 飞书机器人接入
+4. 商业版新增 - 企微客服接入
