+++ projects/app/src/pages/app/detail/components/WorkflowComponents/context.tsx
@@ -48,7 +48,6 @@ import { useRequest2 } from '@fastgpt/web/hooks/useRequest';
 import { formatTime2HM, formatTime2YMDHMW } from '@fastgpt/global/common/string/time';
 import type { InitProps } from '@/pages/app/detail/components/PublishHistoriesSlider';
 import { cloneDeep } from 'lodash';
-import { THelperLine } from '@fastgpt/global/core/workflow/type';
 
 type OnChange<ChangesType> = (changes: ChangesType[]) => void;
 
@@ -135,12 +134,6 @@ type WorkflowContextType = {
   // version history
   historiesDefaultData?: InitProps;
   setHistoriesDefaultData: React.Dispatch<React.SetStateAction<undefined | InitProps>>;
-
-  // helper line
-  helperLineHorizontal?: THelperLine;
-  setHelperLineHorizontal: React.Dispatch<React.SetStateAction<THelperLine | undefined>>;
-  helperLineVertical?: THelperLine;
-  setHelperLineVertical: React.Dispatch<React.SetStateAction<THelperLine | undefined>>;
 
   // chat test
   setWorkflowTestData: React.Dispatch<
@@ -265,14 +258,6 @@ export const WorkflowContext = createContext<WorkflowContextType>({
   saveLabel: '',
   historiesDefaultData: undefined,
   setHistoriesDefaultData: function (value: React.SetStateAction<InitProps | undefined>): void {
-    throw new Error('Function not implemented.');
-  },
-  helperLineHorizontal: undefined,
-  setHelperLineHorizontal: function (value: React.SetStateAction<THelperLine | undefined>): void {
-    throw new Error('Function not implemented.');
-  },
-  helperLineVertical: undefined,
-  setHelperLineVertical: function (value: React.SetStateAction<THelperLine | undefined>): void {
     throw new Error('Function not implemented.');
   },
   getNodeDynamicInputs: function (nodeId: string): FlowNodeInputItemType[] {
@@ -742,11 +727,6 @@ const WorkflowContextProvider = ({
   /* Version histories */
   const [historiesDefaultData, setHistoriesDefaultData] = useState<InitProps>();
 
-  /* helper line */
-  const [helperLineHorizontal, setHelperLineHorizontal] = useState<THelperLine | undefined>(
-    undefined
-  );
-  const [helperLineVertical, setHelperLineVertical] = useState<THelperLine | undefined>(undefined);
   /* event bus */
   useEffect(() => {
     eventBus.on(EventNameEnum.requestWorkflowStore, () => {
@@ -815,12 +795,6 @@ const WorkflowContextProvider = ({
     // version history
     historiesDefaultData,
     setHistoriesDefaultData,
-
-    // helper line
-    helperLineHorizontal,
-    setHelperLineHorizontal,
-    helperLineVertical,
-    setHelperLineVertical,
 
     // chat test
     setWorkflowTestData
